+++ internal/extsvc/bitbucketserver/client.go
@@ -1432,6 +1432,10 @@ func (e *httpError) NoSuchLabelException() bool {
 	return strings.Contains(string(e.Body), bitbucketNoSuchLabelException)
 }
 
+func (e *httpError) MergePreconditionFailedException() bool {
+	return e.StatusCode == 409
+}
+
 const (
 	bitbucketDuplicatePRException       = "com.atlassian.bitbucket.pull.DuplicatePullRequestException"
 	bitbucketNoSuchLabelException       = "com.atlassian.bitbucket.label.NoSuchLabelException"
@@ -1506,4 +1510,37 @@ func (c *Client) CreatePullRequestComment(ctx context.Context, pr *PullRequest, 
 	var resp *Comment
 	_, err := c.send(ctx, "POST", path, qry, &payload, &resp)
 	return err
+}
+
+// ErrNotMergeable is returned by MergePullRequest when the pull request failed
+// to merge, because a precondition is not met.
+var ErrNotMergeable = errors.New("pull request cannot be merged")
+
+func (c *Client) MergePullRequest(ctx context.Context, pr *PullRequest) error {
+	if pr.ToRef.Repository.Slug == "" {
+		return errors.New("repository slug empty")
+	}
+
+	if pr.ToRef.Repository.Project.Key == "" {
+		return errors.New("project key empty")
+	}
+
+	path := fmt.Sprintf(
+		"rest/api/1.0/projects/%s/repos/%s/pull-requests/%d/merge",
+		pr.ToRef.Repository.Project.Key,
+		pr.ToRef.Repository.Slug,
+		pr.ID,
+	)
+
+	qry := url.Values{"version": {strconv.Itoa(pr.Version)}}
+
+	_, err := c.send(ctx, "POST", path, qry, nil, pr)
+	if err != nil {
+		wrappedErr := errors.Unwrap(err)
+		if e, ok := wrappedErr.(*httpError); ok && e.MergePreconditionFailedException() {
+			return errors.Wrap(ErrNotMergeable, err.Error())
+		}
+		return err
+	}
+	return nil
 }